Refactor, add shardsffi to split thing up
