+++ shards/langffi/Cargo.toml
@@ -1,0 +1,24 @@
+[package]
+name = "shards-langffi"
+version = "0.1.0"
+edition = "2021"
+
+[lib]
+crate-type = ["rlib", "staticlib"]
+
+[dependencies]
+shards = { path = "../rust" }
+shards-lang = { path = "../lang" }
+bincode = "1.3.3"
+
+[target.'cfg(unix)'.dependencies]
+libc = "0.2.155"
+
+[target.'cfg(windows)'.dependencies]
+windows = { version = "0.56.0", features = ["Win32_System_Diagnostics_Debug"] }
+
+[build-dependencies]
+cbindgen = "0.24.0"
+
+[profile.dev]
+opt-level = 3
